Move fromAstObject implementation into Object itself.
As long as key parts of the machinery were defined outside the
class, friending Object doesn't provide them access to the
protected constructors.

We also stop using make_shared for the same reason: since it
would be the function directly calling the protected constructors,
it would have to be the friend.  But that's tricky, since the
signature of make_shared is not strictly defined by the standard.
We could guess it pretty well for any C++11 compiler, but it's
probably best not to.