Minor test improvements
test_object.py was trying to call methods that were no longer
publicly available. This passed because it was checking for
errors (just not the ones being raised).
Improve test_channel.py and test_normMap.py to test for a
more explicit exception when using assertRaises.
I checked the usage in other tests and concluded it was best
to leave them as is because in some cases it wasn't clear which
C++ exception would be mapped to the Python exception that was
raised, and in others it was unclear if AST or C++ code would
catch the problem, thus which Python exception would be raised.