Frame.convert and findFrame return a shared_ptr
Change `Frame.convert` and `Frame.findFrame` to return
`std::shared_ptr<FrameSet>` instead of `FrameSet`
and return an empty pointer if the search is unsuccessful
instead of throwing notfound_error. This is more useful for
the user and closer to how AST does it. It also eliminates
the need for notfound_error and avoids raising a C++ exception
for a normal situation.