Add static_cast where needed for assertEqual
In some places assertEqual was called with (size_t, int)
which led to compiler warnings. Fix those with
static_cast<std::size_t> of the int argument.
In one case the size_t argument was cast to int instead,
but I changed it to match the rest of the code.