Swap order of compound mapping and frame methods
Rename `Mapping.of` to `Mapping.then`,
`Mapping.over` to `Mapping.under`
and `Frame.over` to `Frame.under`,
rename the argument from "first" to "next"
and swap the order in which the two mappings or frames
are combined.
This was done to enhance code readability when
mappings are named "xToY", e.g. `aToC = aToB.then(bToC)`
It also matches the order of arguments to `ParallelMap`,
`SeriesMap` and `CmpFrame`.