First cut at some polynomial utilities
These compute polynomial coefficients from simple arrays of
coefficients for 3 common cases: 1->1, 2->1 and 2->2 (nin->nout).
However, the coefficient order is not particularly obvious in the
latter two cases, so I'm holding off merging this.
Note that the unit test for 2->2 does not yet pass
and there is not yet a test for 2->1.