Support FitsChan.getFits<X> to read current card
...if the name is blank, and make the default name blank.
Note: I tried using a pointer to a std::string but that broke
the Python wrapper due to a bug in pybind11 (reported).

Do the same for FitsChan.testFits (a new AST feature).

Enhance test_fitsChan.py to test this new capability.
Also enhance the unit test to test all fitsSet<X> and fitsGet<X>
and to test behavior for a non-existent card.