Add test for FrameDict memory issue
In C++ FrameDict(FrameSet const &) resulted in an unusable
FrameDict, as AST had already deallocated the AST pointer.
The pybind11 wrapper for that constructor did not show the problem,
making the bug difficult to test in Python. Add a function that
shows the bug in Python and call it in test_frameDict.py