audit for ast error-checking
Added assertOK() where it had been left out. This is C code, so
it needs to be done after every astXXX function call, except for
a few (e.g., astAnnul) and where it's in a destructor. Otherwise,
the ast global error state can be set, we don't realise it, and
the error appears to come out of somewhere else.